+++ settings.js
@@ -16,6 +16,7 @@ export const jerryWorkshopTrackerOverlayGui = new Gui();
 export const wormProfitTrackerOverlayGui = new Gui();
 export const magmaCoreProfitTrackerOverlayGui = new Gui();
 export const abandonedQuarryTrackerOverlayGui = new Gui();
+export const archfiendDiceProfitTrackerOverlayGui = new Gui();
 export const fishingProfitTrackerOverlayGui = new Gui();
 
 const categories = ["General", "Chat", "Alerts", "Overlays", "Items and storages", "Rendering", "Commands"]
@@ -542,6 +543,22 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     title: "Alert when a Worm the Fish is caught",
     description: "Shows a title and plays a sound when a Worm the Fish is caught (Dirt Rod fishing).",
     subcategory: "Worm the Fish",
+    value: true
+})
+.addSwitch({
+    category: "Alerts",
+    configName: "alertOnPetLevelUp",
+    title: "Alert when a pet reached max level",
+    description: "Shows a title and plays a sound when a pet reached max level.",
+    subcategory: "Pet",
+    value: true
+})
+.addSwitch({
+    category: "Alerts",
+    configName: "alertOnFishingFestivalEnded",
+    title: "Alert when Fishing Festival is ended",
+    description: "Shows a title and plays a sound when a Fishing Festival is ended. Additionally, shows statistics with the amount of different sharks caught during that Fishing Festival.",
+    subcategory: "Fishing Festival",
     value: true
 })
 .addSwitch({
@@ -1270,6 +1287,45 @@ Example: ${AQUA}/feeshSetRadioactiveVials 5 2024-03-18T14:05:00Z`);
     subcategory: "Abandoned Quarry tracker",
     onClick() {
         ChatLib.command("feeshResetAbandonedQuarry noconfirm", true);
+    }
+})
+
+.addSwitch({
+    category: "Overlays",
+    configName: "archfiendDiceProfitTrackerOverlay",
+    title: "Archfiend Dice profit tracker",
+    description: `Shows an overlay with your Archfiend Dice / High Class Archfiend Dice profits. This overlay has [Session] and [Total] view mode.`,
+    subcategory: "Archfiend Dice profit tracker",
+    value: true
+})
+.addButton({
+    category: "Overlays",
+    configName: "moveArchfiendDiceProfitTrackerOverlay",
+    title: "Move Archfiend Dice profit tracker",
+    description: "Allows to move and resize the overlay text.",
+    subcategory: "Archfiend Dice profit tracker",
+    onClick() {
+        moveOverlay(archfiendDiceProfitTrackerOverlayGui);
+    }
+})
+.addButton({
+    category: "Overlays",
+    configName: "resetArchfiendDiceProfitTrackerSession",
+    title: "Reset Archfiend Dice profit tracker [Session]",
+    description: `Resets tracking for Archfiend Dice profit tracker [Session]. Executes ${AQUA}/feeshResetArchfiendDiceProfit`,
+    subcategory: "Archfiend Dice profit tracker",
+    onClick() {
+        ChatLib.command("feeshResetArchfiendDiceProfit noconfirm", true);
+    }
+})
+.addButton({
+    category: "Overlays",
+    configName: "resetArchfiendDiceProfitTrackerTotal",
+    title: "Reset Archfiend Dice profit tracker [Total]",
+    description: `Resets tracking for Archfiend Dice profit tracker [Total]. Executes ${AQUA}/feeshResetArchfiendDiceProfitTotal`,
+    subcategory: "Archfiend Dice profit tracker",
+    onClick() {
+        ChatLib.command("feeshResetArchfiendDiceProfitTotal noconfirm", true);
     }
 })